+++ pulsar/apps/__init__.py
@@ -76,7 +76,7 @@ import pulsar
 from pulsar import Actor, Monitor, get_actor, EventHandler, async
 from pulsar.utils.structures import OrderedDict
 from pulsar.utils.pep import pickle
-from pulsar.utils.sockets import parse_connection_string, get_connection_string
+from pulsar.utils.internet import parse_connection_string, get_connection_string
 from pulsar.utils.log import LocalMixin
 from pulsar.utils.importer import import_module
 from pulsar.utils.security import gen_unique_id